api/requirements.txt
fastapi==0.115.5
numpy==2.1.3
Pillow==11.0.0
tensorflow==2.18.0
uvicorn==0.32.0
